refactor: Merge webview types with common types and change type name NewTreeNode to TreeNode
