+++ treeAct-view/src/components/markdownRender.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 import Markdown from "markdown-to-jsx";
 import readme from "../../../README.md";
