+++ treeAct-view/src/module/index.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
 import treeReducer from "./tree";
-import { TreeState } from "../types";
+import { TreeState } from "../../../common/types";
 
 export interface RootState {
   treeReducer: TreeState;
