--- treeAct-view/src/types.tsx
@@ -1,9 +1,0 @@
-import { NewTreeNode } from "../../common/types";
-
-export type Name = {
-  id: number;
-  name: string;
-};
-
-// Type of whole tree data
-export type TreeState = { treeData: NewTreeNode[] };
